ptr is long not int

+++ arch/hppa/code.c
@@ -139,7 +139,7 @@ bfcode(struct symtab **a, int cnt)
 	if (cftnsp->stype == STRTY+FTN || cftnsp->stype == UNIONTY+FTN) {
 		/* Function returns struct, adjust arg offset */
 		for (i = 0; i < n; i++)
-			a[i]->soffset += SZPOINT(INT);
+			a[i]->soffset += SZPOINT(LONG);
 	}
 
 	/* recalculate the arg offset and create TEMP moves */
